backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
banks==2.1.3
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
deprecated==1.2.18
dirtyjson==1.0.8
django==5.2.4
djangorestframework==3.16.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
griffe==1.7.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
joblib==1.5.1
llama-cloud==0.1.33
llama-cloud-services==0.6.46
llama-index-core==0.12.48
llama-index-instrumentation==0.2.0
llama-index-workflows==1.1.0
markupsafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
pyyaml==6.0.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
sqlalchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
wrapt==1.17.2
yarl==1.20.1
